custom_sub_v2/pyproject.toml
[project]
name = "custom-sub-v2"
version = "0.1.0"
description = "Article to explainer video tool v2 — OpenRouter direct API, multi-entry Vite, React Web UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "claude-code-sdk>=0.0.20",
    "dashscope>=1.20.0",
    "playwright>=1.48.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "click>=8.1.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "mutagen>=1.47.0",
    "websockets>=13.0",
]

[project.scripts]
custom-sub-v2 = "custom_sub_v2.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/custom_sub_v2"]
